Uninstall Flux image automation controllers
It appears the image automation controllers are using YAML comments to mark places that need to be updated. This is problematic as the same skeleton of apps cannot be used by different Kustomization(s) but different image policies; the main use-case is for the "main" kustomization to specify a "releases" image policy, but other branches to use "sha" policy.