feat(deploy): stop environments with no commits in the last hour
This is meant to conserve resources, and take down environments that are
not being actively worked on. It is done by searching every hour for
such environments, and for one of those add a patch which sets the
"replicas" property to zero on all deployments.

When an environment is deployed to, this patch is removed, which loads
the environment back up.