refactor: simplify deployment workflows
This change translates branch to environment name using the new reusable
action "arikkfir/delivery-env-name", which helps simplify the workflows
as well as add some documentation.

The stale environment garbage collection workflow will now also ensure
that a non-stale environment does not have the staleness patches, as a
precaution measure for cases where the deployment workflow fails for
some reason and the environment stays in the stale state.

[skip ci]