feat(greenstar): use init container instead of job for neo4j migrations
This change changes how we migrate Neo4j data from a Kubernetes Job to
an initialization container (a.k.a. "initContainers") in the backend API
deployment. This ensures that migration happens before a new version of
a backend API container starts, and also makes environment suspension
easier since we don't need to keep a suspended Job object in place.

This required:

1. Updating the greenstar app manifests
2. Updating the stale environments patches
3. Update all existing environments accordingly